[APIView] Expose endpoint for View Approvers for a Language
